pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Warren Whipple <w@warrenwhipple.com>"]
readme = "README.md"

[tool.poetry.scripts]
dev = "app:main"

[tool.poetry.dependencies]
python = "^3.12"
slack-bolt = "^1.18.1"
python-dotenv = "^1.0.1"
openai = "^1.10.0"
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
